move pkg/generator package to rpaas/nginx package
