+++ rpaas/nginx/generate.go
@@ -1,4 +1,4 @@
-package generator
+package nginx
 
 import (
 	"bytes"
