+++ rpaas/nginx/generate_test.go
@@ -1,4 +1,4 @@
-package generator
+package nginx
 
 import (
 	"testing"
